# feat: use sha256-simd