--- a/crypto/hash.go
+++ b/crypto/hash.go
@@ -1,7 +1,7 @@
 package crypto
 
 import (
-	"crypto/sha256"
+	"github.com/minio/sha256-simd"
 )
 
 func Sha256(bytes []byte) []byte {