--- a/statesync/snapshots.go
+++ b/statesync/snapshots.go
@@ -1,17 +1,17 @@
 package statesync
 
 import (
-	"crypto/sha256"
 	"fmt"
 	"math/rand"
 	"sort"
 
+	"github.com/tendermint/tendermint/crypto/tmhash"
 	tmsync "github.com/tendermint/tendermint/libs/sync"
 	"github.com/tendermint/tendermint/p2p"
 )
 
 // snapshotKey is a snapshot key used for lookups.
-type snapshotKey [sha256.Size]byte
+type snapshotKey [tmhash.Size]byte
 
 // snapshot contains data about a snapshot.
 type snapshot struct {
@@ -29,7 +29,7 @@ type snapshot struct {
 // non-deterministic manner. All fields must be equal for the snapshot to be considered the same.
 func (s *snapshot) Key() snapshotKey {
 	// Hash.Write() never returns an error.
-	hasher := sha256.New()
+	hasher := tmhash.New()
 	hasher.Write([]byte(fmt.Sprintf("%v:%v:%v", s.Height, s.Format, s.Chunks)))
 	hasher.Write(s.Hash)
 	hasher.Write(s.Metadata)