Hello,

I wanted to contribute to #272 but noticed that the code changed a lot and some optimizations were rolled back, so I'm not sure if that still makes sense..
Anyway, this is the result on my machine after the changes:
```
goos: linux
goarch: amd64
pkg: github.com/tendermint/tendermint/crypto/merkle
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                             │  original   │            sha256-simd2             │
                             │   sec/op    │   sec/op     vs base                │
HashAlternatives/recursive-8   64.21µ ± 1%   28.46µ ± 2%  -55.67% (p=0.000 n=10)
HashAlternatives/iterative-8   64.12µ ± 1%   28.79µ ± 1%  -55.10% (p=0.000 n=10)
geomean                        64.17µ        28.63µ       -55.39%

                             │   original   │             sha256-simd2              │
                             │     B/op     │     B/op      vs base                 │
HashAlternatives/recursive-8   24.83Ki ± 0%   24.83Ki ± 0%       ~ (p=1.000 n=10) ¹
HashAlternatives/iterative-8   27.45Ki ± 0%   27.45Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                        26.11Ki        26.11Ki       +0.00%
¹ all samples are equal

                             │  original  │            sha256-simd2             │
                             │ allocs/op  │ allocs/op   vs base                 │
HashAlternatives/recursive-8   497.0 ± 0%   497.0 ± 0%       ~ (p=1.000 n=10) ¹
HashAlternatives/iterative-8   498.0 ± 0%   498.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                        497.5        497.5       +0.00%
¹ all samples are equal
```

Note that 11800h has AVX-512 and SHA extensions, so the speedup shouldn't be the same everywhere but I still hope that improves things for anything with SHA extensions, like cheap Celeron boxes or ARM boards.

---

Remaining questions:
- I see some optimizations like #351 are on the deprecated branch but not here, is porting them something you want?
- Should I change crypto/sha256 to sha256-simd everywhere?