@evan-forbes 2023-01-30
> Remaining questions:
I see some optimizations like https://github.com/celestiaorg/celestia-core/pull/351 are on the deprecated branch but not here, is porting them something you want?

oh shoot! good point, we also totally forgot about that. Yes, we should be able to port that change, if it hasn't been upstreamed already (I can check on that)

> Should I change crypto/sha256 to sha256-simd everywhere?

yeah definitely, but we'll have to go through our other repos and make sure we are referring to the consts defined here. 

there's one failing test in the mempool that is probably just due to something silly that I'll also checkout tmrw or the next day
```
--- FAIL: TestTxMempool_ExpiredTxs_Timestamp (0.02s)
    mempool_test.go:567: Transaction 09BC3C005351A6DEE71D2F8CC06FC8E5C010F81A10E805F26EADF3674FEB397C should still be in the mempool, but is not
    mempool_test.go:567: Transaction B7240B89DEBE709CC3FBFE1C41854C0C555B6C516B9E77AD64B8BE2ECF4CE7B9 should still be in the mempool, but is not
    mempool_test.go:567: Transaction 6756C4752A75DD57E02D23D81A7CA4ECD80F74036506827F1C46D101ABEC2C4A should still be in the mempool, but is not
    mempool_test.go:567: Transaction E734AD200FC78ECEEFD69C3CF2C9BAAF00FD107B921FCDF3632C72AFCDCDAD71 should still be in the mempool, but is not
    mempool_test.go:567: Transaction 0EDDA5067BA45376EA540296A615FF273C9371405262B8F517A9C07A404C0FD3 should still be in the mempool, but is not
    mempool_test.go:567: Transaction 9784E8F44F9AA02E76CCC17BB64CB1D6946AC5189360F6CA7B5A3201578B447D should still be in the mempool, but is not
FAIL
coverage: 84.5% of statements
```