@shyba 2023-01-30
Thanks for the quick reply!

> which should be pointing to the tmhash.New function, and not just sha256. That will actually provide the most benefit, since we are using that to calculate the data root

Changed pkg/consts

> yeah definitely, but we'll have to go through our other repos and make sure we are referring to the consts defined here.

Nice! Will do that.
For the local matches, I wonder if they should all refer to crypto/tmhash?
There is also a Sha256 function in crypto/hash

Also, I was thinking in leaving test modules untouched as they aren't performance critical and this could help checking if something broke.. Maybe on a case by case

> there's one failing test in the mempool that is probably just due to something silly that I'll also checkout tmrw or the next day

Cool, I was going to fix here but it pass locally.. I will try to clear cache and comment if I find anything