@shyba 2023-01-31
Replaced everywhere to use tmhash instead of crypto/sha256, but did that on a separate commit because after a while I wasn't sure if that was the goal..
It didn't seem like a good idea for readability, specially for crypto code like secp256k1 internals, so if you want that in another way I can revert or maybe alias.