@shyba 2023-01-31
Since I was skeptical of the results, I re-ran them on the target machine which will run my nodes. It is a Celeron N5100 (6w lol)

It lacks even basic AVX but does have SHA extensions. I was expecting some improvement but definitely not 72%!
```
goos: linux
goarch: amd64
pkg: github.com/tendermint/tendermint/crypto/merkle
cpu: Intel(R) Celeron(R) N5100 @ 1.10GHz
                             │   original   │                 simd                 │                reuse                │
                             │    sec/op    │    sec/op     vs base                │   sec/op     vs base                │
HashAlternatives/recursive-4   428.81µ ± 0%   118.07µ ± 1%  -72.47% (p=0.000 n=10)   96.14µ ± 0%  -77.58% (p=0.000 n=10)
HashAlternatives/iterative-4   432.93µ ± 0%   116.41µ ± 1%  -73.11% (p=0.000 n=10)   96.84µ ± 0%  -77.63% (p=0.000 n=10)
geomean                         430.9µ         117.2µ       -72.79%                  96.49µ       -77.61%

                             │   original    │                  simd                  │                reuse                 │
                             │     B/op      │     B/op       vs base                 │     B/op      vs base                │
HashAlternatives/recursive-4   24.828Ki ± 0%   24.828Ki ± 0%       ~ (p=1.000 n=10) ¹   6.522Ki ± 0%  -73.73% (p=0.000 n=10)
HashAlternatives/iterative-4   27.453Ki ± 0%   27.453Ki ± 0%       ~ (p=1.000 n=10) ¹   9.147Ki ± 0%  -66.68% (p=0.000 n=10)
geomean                         26.11Ki         26.11Ki       +0.00%                    7.724Ki       -70.41%
¹ all samples are equal

                             │  original  │                simd                 │               reuse                │
                             │ allocs/op  │ allocs/op   vs base                 │ allocs/op   vs base                │
HashAlternatives/recursive-4   497.0 ± 0%   497.0 ± 0%       ~ (p=1.000 n=10) ¹   399.0 ± 0%  -19.72% (p=0.000 n=10)
HashAlternatives/iterative-4   498.0 ± 0%   498.0 ± 0%       ~ (p=1.000 n=10) ¹   400.0 ± 0%  -19.68% (p=0.000 n=10)
geomean                        497.5        497.5       +0.00%                    399.5       -19.70%
¹ all samples are equal
```

`original` is the current code
`simd` is this PR
`reuse` is a branch where I adapted the idea of #351 

That adaptation lives [on this separate commit](https://github.com/contribforks/celestia-core/commit/71f933375ddaa2ec062f849cbb89b71972462e72). I can open a PR if you want.