@evan-forbes 2023-05-26
closing due to us planning on using the std lib in go 1.21 https://github.com/celestiaorg/celestia-app/issues/1826

thanks for doing this @shyba :purple_heart:, your contribution was still valuable although we did not end up merging. we should have more time in the future to spend on optimizations